src/worker/requirements.txt
telethon==1.37.0
redis==5.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
